requirements.txt
docker
python-dotenv
paramiko
schedule
loguru
